# src/ControlPlane/Bagatka.AiSloth.WebApi/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "Bagatka.AiSloth.WebApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/instructions": {
      "put": {
        "tags": [
          "Chats"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstructionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/workspaces/{id}": {
      "get": {
        "tags": [
          "Workspaces"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSummary"
                }
              }
            }
          }
        }
      }
    },
...
    "/workspaces/{workspaceId}/chats": {
      "post": {
        "tags": [
          "Chats"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSummary"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Chats"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfChatSummary"
                }
              }
            }
          }
        }
      }
    },
    "/chats/{id}": {
      "get": {
        "tags": [
          "Chats"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSummary"
                }
              }
            }
          }
        }
      }
    },
    "/chats/{id}/messages": {
      "post": {
        "tags": [
          "Chats"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessage"
                }
              }
            }
          }
        }
      }
    },
    "/chats/{id}/stop": {
      "post": {
        "tags": [
          "Chats"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/chats/{id}/events": {
      "get": {
        "tags": [
          "Chats"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "required": [
                    "data"
                  ],
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ChatEventData"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/chats/{id}/push": {
      "post": {
        "tags": [
          "Chats"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PushedSource"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/harness-state": {
      "get": {
        "tags": [
          "Chats"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HarnessStateSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/harness-state/{harness}": {
      "delete": {
        "tags": [
          "Chats"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "harness",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shared",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/instructions": {
      "get": {
        "tags": [
          "Chats"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Instructions"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Chats"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstructionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AccessLevel": {
        "enum": [
          "Read",
          "Write",
          "Manage"
        ],
        "description": "How much a person may do with a resource. Each level includes the ones below it, so compare\nlevels: \u0060access \u0026gt;= AccessLevel.Write\u0060. What a level allows for its own data, each module\ndecides."
      },
...
      "AgentAccountId": {
        "description": "Identifies an agent account."
      },
...
      "ChatEventData": {
        "required": [
          "sequence",
          "at",
          "event"
        ],
        "type": "object",
        "properties": {
          "sequence": {
            "type": "integer",
            "format": "int64"
          },
          "at": {
            "type": "string",
            "format": "date-time"
          },
          "event": {
            "$ref": "#/components/schemas/ChatEventBody"
          }
        }
      },
      "ChatId": {
        "description": "Identifies a chat."
      },
      "ChatMessage": {
        "required": [
          "id",
          "chatId",
          "sentBy",
          "text",
          "sentAt",
          "isProposal"
        ],
        "type": "object",
        "properties": {
          "id": {
            "description": "The message.",
            "$ref": "#/components/schemas/MessageId"
          },
          "chatId": {
            "description": "Its chat.",
            "$ref": "#/components/schemas/ChatId"
          },
          "sentBy": {
            "description": "Who sent it.",
            "$ref": "#/components/schemas/UserId"
          },
          "text": {
            "type": "string",
            "description": "What it says."
          },
          "sentAt": {
            "type": "string",
            "description": "When it was sent.",
            "format": "date-time"
          },
          "isProposal": {
            "type": "boolean",
            "description": "Whether it is a proposal: it never reaches the agent, and the account\u0027s owner may send it on."
          }
        },
        "description": "A message someone sent to a chat."
      },
      "ChatSummary": {
        "required": [
          "id",
          "nookId",
          "workspaceId",
          "startedBy",
          "startedAt",
          "working",
          "harness",
          "account",
          "accountOwner"
        ],
        "type": "object",
        "properties": {
          "id": {
            "description": "The chat.",
            "$ref": "#/components/schemas/ChatId"
          },
          "nookId": {
            "description": "The nook its agent works in.",
            "$ref": "#/components/schemas/NookId"
          },
          "workspaceId": {
            "description": "The nook\u0027s workspace.",
            "$ref": "#/components/schemas/WorkspaceId"
          },
          "startedBy": {
            "description": "Who started it.",
            "$ref": "#/components/schemas/UserId"
          },
          "startedAt": {
            "type": "string",
            "description": "When it started.",
            "format": "date-time"
          },
          "working": {
            "type": "boolean",
            "description": "Whether the agent is working on a turn, or messages wait for it."
          },
          "harness": {
            "type": "string",
            "description": "The harness that runs its agent."
          },
          "account": {
            "description": "The agent account that pays for its work.",
            "$ref": "#/components/schemas/AgentAccountId"
          },
          "accountOwner": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "The owner of the personal account it runs on, whose messages alone reach the agent: everyone\nelse\u0027s are proposals. \u0060null\u0060 on the workspace\u0027s account, which people with Write\non the workspace use.",
                "$ref": "#/components/schemas/UserId"
              }
            ]
          }
        },
        "description": "A chat as the people with access to its nook see it."
      },
...
      "HarnessStateSummary": {
        "required": [
          "harness",
          "shared",
          "savedAt",
          "bytes",
          "savedFrom"
        ],
        "type": "object",
        "properties": {
          "harness": {
            "type": "string",
            "description": "The harness, such as \u0060claude-code\u0060."
          },
          "shared": {
            "type": "boolean",
            "description": "Whether it is the workspace\u0027s, kept from the chats on the workspace\u0027s accounts and given to all of\nthem, rather than the actor\u0027s own, kept from the chats on their own accounts."
          },
          "savedAt": {
            "type": "string",
            "description": "When a chat last saved it.",
            "format": "date-time"
          },
          "bytes": {
            "type": "integer",
            "description": "The size of its history.",
            "format": "int64"
          },
          "savedFrom": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "The chat that saved it last, unless that chat is gone.",
                "$ref": "#/components/schemas/ChatId"
              }
            ]
          }
        },
        "description": "Harness state for one harness, as the list of a workspace\u0027s harness state shows it."
      },
...
      "Instructions": {
        "required": [
          "workspace",
          "personal"
        ],
        "type": "object",
        "properties": {
          "workspace": {
            "type": "string",
            "description": "The workspace\u0027s, for everyone\u0027s chats there."
          },
          "personal": {
            "type": "string",
            "description": "The person\u0027s own, for the chats they start."
          }
        },
        "description": "What every agent of a person\u0027s chats in a workspace is told to follow, besides each repository\u0027s\nown instructions: the workspace\u0027s first, then the person\u0027s own. Empty for none."
      },
      "InstructionsRequest": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          }
        }
      },
...
      "NookId": {
        "description": "Identifies a nook."
      },
      "NookRepository": {
        "required": [
          "repository"
        ],
        "type": "object",
        "properties": {
          "repository": {
            "description": "One of the workspace\u0027s repositories (\u0060ISourcesApi\u0060).",
            "$ref": "#/components/schemas/RepositoryId"
          },
          "branch": {
            "type": [
              "null",
              "string"
            ],
            "description": "The branch to start from, or \u0060null\u0060 for its default branch."
          }
        },
        "description": "A repository a new nook starts with."
      },
...
      "PageOfChatSummary": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatSummary"
            },
            "description": "The items, in the list\u0027s stable order."
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ],
            "description": "Pass it in the next PageRequest; \u0060null\u0060 on the last page."
          }
        },
        "description": "One page of a keyset-paginated list."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PushedSource": {
        "required": [
          "source",
          "branch",
          "commits",
          "branchUrl",
          "pullRequestUrl",
          "problem"
        ],
        "type": "object",
        "properties": {
          "source": {
            "type": "string"
          },
          "branch": {
            "type": [
              "null",
              "string"
            ]
          },
          "commits": {
            "type": "integer",
            "format": "int32"
          },
          "branchUrl": {
            "type": [
              "null",
              "string"
            ],
            "format": "uri"
          },
          "pullRequestUrl": {
            "type": [
              "null",
              "string"
            ],
            "format": "uri"
          },
          "problem": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PushRequest": {
        "type": "object",
        "properties": {
          "sources": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "branch": {
            "type": [
              "null",
              "string"
            ]
          },
          "pullRequest": {
            "type": "boolean",
            "default": false
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SendMessageRequest": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "proposal": {}
        }
      },
...
      "StartChatRequest": {
        "required": [
          "provider",
          "harness",
          "account"
        ],
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "harness": {
            "type": "string"
          },
          "account": {
            "$ref": "#/components/schemas/AgentAccountId"
          },
          "repositories": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/NookRepository"
            }
          },
          "copyOf": {},
          "checkpoint": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
...
      "UserId": {
        "description": "Identifies a user. It lives in Foundation, not in a module, because Actor needs it."
      },
...
      "WorkspaceId": {},
      "WorkspaceSummary": {
        "required": [
          "id",
          "name",
          "access"
        ],
        "type": "object",
        "properties": {
          "id": {
            "description": "The workspace.",
            "$ref": "#/components/schemas/WorkspaceId"
          },
          "name": {
            "type": "string",
            "description": "Its display name."
          },
          "access": {
            "description": "How much the caller may do in it.",
            "$ref": "#/components/schemas/AccessLevel"
          }
        },
        "description": "A workspace as one person with access sees it."
      }
    }
  },
...
}